separate SlackChannel interface, update links to new location
