+++ src/services/conversion.ts
@@ -1,7 +1,7 @@
 import { ICohort } from '../models/Cohort';
 import User, { IUser } from '../models/User';
 import { isCohortSubject } from '../utils/typeGuards';
-import { SlackChannel } from '../utils/slack/channel';
+import { SlackChannel } from '../slack/types/channel';
 import { SlackMember } from '../slack/types/member';
 
 class ConvesionService {
