+++ src/slack/types/channel.ts
@@ -1,0 +1,8 @@
+export interface SlackChannel {
+  slackId: string;
+  name: string;
+  type: string;
+  numberOfMembers: number;
+  startDate?: string;
+  endDate?: string;
+}
